data output test

+++ ecr-repository/services/prediction_service.py
@@ -27,7 +27,7 @@ class PredictionService:
         return os.environ.get('PREDICT_QUEUE')
         
     def _inv_softmax(self, x):
-        return tf.convert_to_tensor((K.log(x) + K.log(math.log(10.))))
+        return tf.convert_to_tensor((K.log(x) + K.log(math.log(10.))), dtype=tf.float32)
         # .numpy()[0]
 
     def _convert_output(self, predict_value):
